backend/requirements.txt
fastapi
uvicorn
motor
pydantic[email]
python-jose[cryptography]
pydantic-settings
passlib[bcrypt]
bcrypt
python-dotenv
google-generativeai
slowapi
python-multipart
